moved company logic to CompaniesController
